requirements.txt
aiogram~=3.3.0
SQLAlchemy~=2.0.29
pydantic~=2.9.2
alembic~=1.13.1
asyncpg
aiogram_dialog
uvloop
aiodns
pydantic-settings
psycopg2-binary

python-dotenv~=1.0.1